chore:#7 변수명 변경
